backend/pyproject.toml

[tool.pdm]
distribution = false
[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "lsprdev", email = "github@lspr.dev"},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

